Light Year and Mile Conversion
This is an program which converts light years to miles and vice-versa. I've noticed that it is difficult for people to picture an object being light years away (I'll admit I've had that problem in the past), but now it can be easier since it the light year distance can be presented in miles.